uploded dataset
This is a dataset called "Iris dataset". It consists of 150 rows and 6 columns. The first 5 columns contain measurements of the sepal length, sepal width, petal length, and petal width of three different species of iris flowers. The last column contains the species of the iris. There are 50 observations for each of the three species: Iris-setosa, Iris-versicolor, and Iris-virginica.

This dataset is often used in machine learning and data science for tasks such as classification and clustering.